add injection of pathsMap in QueryBuilder
