--- src/main/java/ru/alcereo/entities/ProcessorsVersionsEntity.java
@@ -1,82 +1,0 @@
-package ru.alcereo.entities;
-
-import org.hibernate.annotations.Filter;
-
-import javax.persistence.*;
-import java.util.Set;
-
-/**
- * Created by alcereo on 15.04.17.
- */
-@Entity
-@Table(name = "processors_versions"
-//        schema = "public",
-//        catalog = "TestDB"
-)
-public class ProcessorsVersionsEntity {
-    private int id;
-    private String name;
-    private Set<ProcessorsEntity> processorsUsed;
-
-    @Id
-    @Column(name = "id", nullable = false)
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "name", nullable = true, length = 64)
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        ProcessorsVersionsEntity that = (ProcessorsVersionsEntity) o;
-
-        if (id != that.id) return false;
-        if (name != null ? !name.equals(that.name) : that.name != null) return false;
-
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        int result = id;
-        result = 31 * result + (name != null ? name.hashCode() : 0);
-        return result;
-    }
-
-    @OneToMany(mappedBy = "processorVersion")
-    public Set<ProcessorsEntity> getProcessorsUsed() {
-        return processorsUsed;
-    }
-
-    public void setProcessorsUsed(Set<ProcessorsEntity> processorsUsed) {
-        this.processorsUsed = processorsUsed;
-    }
-
-    @PostLoad
-    void postLoad() {
-        System.out.println("postLoad");
-    }
-
-    @Override
-    public String toString() {
-        return "ProcessorsVersionsEntity{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
